requirements.txt
# Core dependencies
telethon==1.27.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0

# For dbt
dbt-core==1.5.0
dbt-postgres==1.5.0

# Logging
structlog==24.1.0
